move the dict to a new file
